Changement de version pour mettre a jour la version qui a été rendu

+++ virtual_machine/srcs/ft_processus_creation.c
@@ -6,7 +6,7 @@
 /*   By: paul <paul@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/03 10:02:44 by paul              #+#    #+#             */
-/*   Updated: 2020/06/04 17:28:53 by paul             ###   ########.fr       */
+/*   Updated: 2020/06/14 18:31:44 by paul             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ t_process		*ft_processus_cpy(t_process *processus, size_t pc, size_t no)
 	if (!(new_processus = ft_processus_create(no, pc)))
 		return (NULL);
 	ft_memcpy(&(new_processus->registre), &(processus->registre),
-	16 * sizeof(int));
+	16 * sizeof(long));
 	new_processus->carry = processus->carry;
 	return (new_processus);
 }
